fix: generalize multi action button
